# packages/contracts/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Spott API
  version: 1.0.0
  description: Spott iOS/Web 单一契约。具备 Idempotency-Key 参数的副作用写入由调用方提供稳定幂等键； 媒体上传、完成、附件与举报均以调用方保留的 UUID 和服务端持久回执支持响应丢失恢复； 业务状态始终由服务端裁决。
  contact:
    name: Spott Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.spott.jp/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /media/upload-intents:
    post:
      operationId: createMediaUploadIntent
      summary: Create or recover a durable image upload intent
      tags:
        - Media
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaUploadInput'
      responses:
        '200':
          description: Eligible pending replay with a fresh capability, or authoritative state/receipt only
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Pragma:
              schema:
                type: string
                const: no-cache
            Referrer-Policy:
              schema:
                type: string
                const: no-referrer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadReplayResponse'
        '201':
          description: New pending upload with a same-origin gateway capability
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Pragma:
              schema:
                type: string
                const: no-cache
            Referrer-Policy:
              schema:
                type: string
                const: no-referrer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadCapabilityResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/upload-attempts/{attemptId}:
    get:
      operationId: recoverMediaUploadAttempt
      summary: Recover one owner-scoped upload attempt after response loss
      tags:
        - Media
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Eligible recovery capability or authoritative in-progress/committed state
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Pragma:
              schema:
                type: string
                const: no-cache
            Referrer-Policy:
              schema:
                type: string
                const: no-referrer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadReplayResponse'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/upload-attempts/{attemptId}/content:
    put:
      operationId: uploadMediaAttemptContent
      summary: Stream exact bytes through the capability-only Spott upload gateway
      description: Cookies and Authorization are forbidden. The capability binds this path, MIME, byte length, SHA-256, owner, asset, generation and expiry. The server enforces one total inbound deadline.
      tags:
        - Media
      security: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: X-Spott-Upload-Capability
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 4096
          description: Opaque Spott gateway capability.
        - name: Content-Type
          in: header
          required: true
          schema:
            type: string
            enum:
              - image/jpeg
              - image/png
              - image/webp
              - image/heic
        - name: Content-Length
          in: header
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 20971520
        - name: X-Content-SHA256
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/SHA256Input'
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
          image/heic:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Authoritative in-progress or committed gateway outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaGatewayUploadResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '408':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '415':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/{id}/complete:
    post:
      operationId: completeMediaUpload
      summary: Complete an upload after verifying its content hash
      tags:
        - Media
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: X-Content-SHA256
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/SHA256Input'
      responses:
        '200':
          description: Authoritative processing state; exact replay never queues duplicate work
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaCompletionResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/{id}:
    delete:
      operationId: abandonMediaAsset
      summary: Idempotently abandon an owned unreferenced media asset
      tags:
        - Media
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Authoritative abandonment and cleanup state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAbandonResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/{id}/attach/event/{eventId}:
    post:
      operationId: attachEventMedia
      summary: Attach one of at most six processed images to an event
      tags:
        - Media
        - Events
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: eventId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventMediaAttachmentInput'
      responses:
        '200':
          description: Exact committed event-media attachment envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMediaAttachment'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/{id}/attach/profile:
    post:
      operationId: attachProfileAvatar
      summary: Replace the current user's avatar with a ready approved profile asset
      tags:
        - Media
        - Profiles
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Exact committed avatar attachment envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileMediaAttachment'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/{id}/attach/group/{groupId}:
    post:
      operationId: attachGroupCover
      summary: Replace a group cover with a ready approved asset owned by the group owner
      tags:
        - Media
        - Groups
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: groupId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Exact committed group-cover attachment envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMediaAttachment'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /media/events/{eventId}/arrangement:
    post:
      operationId: arrangeEventMedia
      summary: Idempotently replace the order of an event's existing media assets
      tags:
        - Media
        - Events
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventMediaArrangementInput'
      responses:
        '200':
          description: Exact committed arrangement envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMediaArrangement'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OpsSession:
      type: apiKey
      in: cookie
      name: __Host-spott_ops
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        oneOf:
          - $ref: '#/components/schemas/UUID'
          - type: string
            maxLength: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  responses:
    Error:
      description: Stable business error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    SHA256:
      type: string
      pattern: ^[a-f0-9]{64}$
      description: Canonical lowercase hexadecimal SHA-256.
    SHA256Input:
      type: string
      pattern: ^[A-Fa-f0-9]{64}$
      description: Case-insensitive hexadecimal SHA-256 input; the server canonicalizes it to lowercase.
    MediaUploadInput:
      type: object
      additionalProperties: false
      required:
        - purpose
        - filename
        - mimeType
        - byteSize
        - contentSha256
      properties:
        purpose:
          type: string
          enum:
            - event_cover
            - profile_avatar
            - group_cover
            - report_evidence
            - share_poster
        filename:
          type: string
          minLength: 1
          maxLength: 255
        mimeType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
            - image/heic
        byteSize:
          type: integer
          minimum: 1
          maximum: 20971520
        focalX:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
        focalY:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
        contentSha256:
          $ref: '#/components/schemas/SHA256Input'
    MediaUploadRequiredHeaders:
      type: object
      additionalProperties: false
      required:
        - Content-Type
        - Content-Length
        - X-Content-SHA256
        - X-Spott-Upload-Capability
      properties:
        Content-Type:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
            - image/heic
        Content-Length:
          type: string
          pattern: ^[1-9][0-9]*$
        X-Content-SHA256:
          $ref: '#/components/schemas/SHA256'
        X-Spott-Upload-Capability:
          type: string
          minLength: 1
          maxLength: 4096
    MediaCommittedUploadReceipt:
      type: object
      additionalProperties: false
      required:
        - assetId
        - state
        - leaseState
        - committedAt
      properties:
        assetId:
          $ref: '#/components/schemas/UUID'
        state:
          type: string
          const: committed
        leaseState:
          type: string
          const: committed
        committedAt:
          type: string
          format: date-time
    MediaUploadCapabilityResponse:
      type: object
      additionalProperties: false
      required:
        - attemptId
        - assetId
        - state
        - uploadUrl
        - capability
        - requiredHeaders
        - expiresAt
        - maxBytes
        - requestDeadlineMs
      properties:
        attemptId:
          $ref: '#/components/schemas/UUID'
        assetId:
          $ref: '#/components/schemas/UUID'
        state:
          type: string
          const: pending_upload
        uploadUrl:
          type: string
          format: uri
          description: Same-origin Spott gateway URL; never a provider URL.
        capability:
          type: string
          minLength: 1
          maxLength: 4096
        requiredHeaders:
          $ref: '#/components/schemas/MediaUploadRequiredHeaders'
        expiresAt:
          type: string
          format: date-time
        maxBytes:
          type: integer
          minimum: 1
          maximum: 20971520
        requestDeadlineMs:
          type: integer
          minimum: 1
    MediaUploadUncommittedStateResponse:
      type: object
      additionalProperties: false
      required:
        - attemptId
        - assetId
        - state
        - leaseState
      properties:
        attemptId:
          $ref: '#/components/schemas/UUID'
        assetId:
          $ref: '#/components/schemas/UUID'
        state:
          type: string
          enum:
            - pending_upload
            - abandoned
            - deleted
        leaseState:
          type: string
          enum:
            - idle
            - receiving
            - provider_writing
            - in_progress
            - failed_cleanup_pending
            - failed_clean
    MediaUploadCommittedStateResponse:
      type: object
      additionalProperties: false
      required:
        - attemptId
        - assetId
        - state
        - leaseState
        - receipt
      properties:
        attemptId:
          $ref: '#/components/schemas/UUID'
        assetId:
          $ref: '#/components/schemas/UUID'
        state:
          type: string
          enum:
            - committed
            - uploaded
            - processing
            - ready
            - rejected
            - abandoned
            - deleted
        leaseState:
          type: string
          const: committed
        receipt:
          $ref: '#/components/schemas/MediaCommittedUploadReceipt'
    MediaUploadStateResponse:
      oneOf:
        - $ref: '#/components/schemas/MediaUploadUncommittedStateResponse'
        - $ref: '#/components/schemas/MediaUploadCommittedStateResponse'
    MediaUploadReplayResponse:
      oneOf:
        - $ref: '#/components/schemas/MediaUploadCapabilityResponse'
        - $ref: '#/components/schemas/MediaUploadStateResponse'
    MediaGatewayInProgressResponse:
      type: object
      additionalProperties: false
      required:
        - attemptId
        - assetId
        - state
        - leaseState
      properties:
        attemptId:
          $ref: '#/components/schemas/UUID'
        assetId:
          $ref: '#/components/schemas/UUID'
        state:
          type: string
          const: pending_upload
        leaseState:
          type: string
          enum:
            - receiving
            - provider_writing
            - in_progress
    MediaGatewayUploadResponse:
      oneOf:
        - $ref: '#/components/schemas/MediaGatewayInProgressResponse'
        - $ref: '#/components/schemas/MediaUploadCommittedStateResponse'
    MediaCompletionResponse:
      type: object
      additionalProperties: false
      required:
        - assetId
        - state
        - moderationState
      properties:
        assetId:
          $ref: '#/components/schemas/UUID'
        state:
          type: string
          enum:
            - uploaded
            - processing
            - ready
            - rejected
        moderationState:
          type: string
          enum:
            - pending
            - approved
            - rejected
    MediaAbandonResponse:
      type: object
      additionalProperties: false
      required:
        - assetId
        - state
        - cleanupState
      properties:
        assetId:
          $ref: '#/components/schemas/UUID'
        state:
          type: string
          enum:
            - abandoned
            - deleted
        cleanupState:
          type: string
          enum:
            - not_required
            - queued
            - waiting_for_fence
            - complete
    EventMediaAttachmentInput:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          enum:
            - cover
            - gallery
          default: cover
        sortOrder:
          type: integer
          minimum: 0
          maximum: 5
          default: 0
    EventMediaAttachment:
      type: object
      additionalProperties: false
      required:
        - id
        - eventId
        - assetId
        - kind
        - sortOrder
        - mediaCount
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        eventId:
          $ref: '#/components/schemas/UUID'
        assetId:
          $ref: '#/components/schemas/UUID'
        kind:
          type: string
          enum:
            - cover
            - gallery
        sortOrder:
          type: integer
          minimum: 0
          maximum: 5
        mediaCount:
          type: integer
          minimum: 1
          maximum: 6
    ProfileMediaAttachment:
      type: object
      additionalProperties: false
      required:
        - assetId
        - profileId
        - url
        - version
      properties:
        assetId:
          $ref: '#/components/schemas/UUID'
        profileId:
          $ref: '#/components/schemas/UUID'
        url:
          type: string
          format: uri
        version:
          type: integer
          format: int64
          minimum: 1
    GroupMediaAttachment:
      type: object
      additionalProperties: false
      required:
        - assetId
        - groupId
        - url
        - version
      properties:
        assetId:
          $ref: '#/components/schemas/UUID'
        groupId:
          $ref: '#/components/schemas/UUID'
        url:
          type: string
          format: uri
        version:
          type: integer
          format: int64
          minimum: 1
    EventMediaArrangementInput:
      type: object
      additionalProperties: false
      required:
        - orderedAssetIds
      properties:
        orderedAssetIds:
          type: array
          minItems: 1
          maxItems: 6
          uniqueItems: true
          items:
            $ref: '#/components/schemas/UUID'
    EventMediaArrangement:
      type: object
      additionalProperties: false
      required:
        - eventId
        - orderedAssetIds
        - version
      properties:
        eventId:
          $ref: '#/components/schemas/UUID'
        orderedAssetIds:
          type: array
          minItems: 1
          maxItems: 6
          uniqueItems: true
          items:
            $ref: '#/components/schemas/UUID'
        version:
          type: integer
          format: int64
          minimum: 1
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
            - retryable
            - fieldErrors
            - actions
            - meta
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            retryable:
              type: boolean
            fieldErrors:
              type: array
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                  message:
                    type: string
            actions:
              type: array
              items:
                type: object
                required:
                  - type
                  - label
                properties:
                  type:
                    type: string
                  label:
                    type: string
            meta:
              type: object
              additionalProperties: true